优化编码器初始化，调整编码器获取值函数参数类型，修正电机控制逻辑

+++ hareware/encoder.h
@@ -5,5 +5,5 @@
 
 void Encoder_Init(void);
 
-u8 Encoder_GetValue(uint8_t encoder);
+int Encoder_GetValue(int encoder);
 #endif // !EN
